+++ ws/packet/packet.go
@@ -4,7 +4,7 @@ import "encoding/json"
 
 type Packet struct {
 	Opcode int `json:"op,omitempty"`
-	Sequence int `json:"s,omitempty"`
+	Sequence int64 `json:"s,omitempty"`
 	Event string `json:"t,omitempty"`
 	Data interface{} `json:"d"`
 }
